+++ src/librustc/middle/ty.rs
@@ -98,7 +98,7 @@ pub struct mt {
 
 #[deriving(Eq, Encodable, Decodable)]
 pub enum vstore {
-    vstore_fixed(uint),
+    vstore_fixed(u64),
     vstore_uniq,
     vstore_box,
     vstore_slice(Region)
@@ -4453,16 +4453,16 @@ pub fn normalize_ty(cx: ctxt, t: t) -> t {
 }
 
 // Returns the repeat count for a repeating vector expression.
-pub fn eval_repeat_count(tcx: ctxt, count_expr: @ast::expr) -> uint {
+pub fn eval_repeat_count(tcx: ctxt, count_expr: @ast::expr) -> u64 {
     match const_eval::eval_const_expr_partial(tcx, count_expr) {
       Ok(ref const_val) => match *const_val {
-        const_eval::const_int(count) => return count as uint,
-        const_eval::const_uint(count) => return count as uint,
+        const_eval::const_int(count) => return count as u64,
+        const_eval::const_uint(count) => return count,
         const_eval::const_float(count) => {
             tcx.sess.span_err(count_expr.span,
                               "expected signed or unsigned integer for \
                                repeat count but found float");
-            return count as uint;
+            return count as u64;
         }
         const_eval::const_str(_) => {
             tcx.sess.span_err(count_expr.span,
